refactor: change profileApi
