+++ assets/api/instance.ts
@@ -1,0 +1,6 @@
+import axios from 'axios'
+
+export const instance = axios.create({
+  baseURL: 'https://inctagram-api.vercel.app/api',
+  withCredentials: true,
+})
