+++ assets/api/user/profileApi.ts
@@ -1,48 +1,10 @@
-import { baseUrl } from '@/assets/api/common.api'
+import { AxiosResponse } from 'axios'
+import { instance } from '../instance'
 import { userRouts } from '@/components/common/User/userRouts'
-import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
-import { ProfileData, UpdateProfileData } from './userTypes'
-import { ServerErrorResponse } from '../auth/authTypes'
+import { ProfileData } from './userTypes'
 
-export const profileApi = createApi({
-  reducerPath: 'profileApi',
-  baseQuery: fetchBaseQuery({
-    baseUrl,
-    credentials: 'include',
-  }),
-  endpoints: (builder) => ({
-    getProfile: builder.query<ProfileData, void>({
-      query: () => ({
-        url: userRouts.profile,
-        method: 'GET',
-        headers: {
-          Authorization: `Bearer ${localStorage.getItem('accessToken')}`,
-        },
-      }),
-    }),
-    updateProfile: builder.mutation<void | ServerErrorResponse, UpdateProfileData>({
-      query: (userData) => ({
-        url: userRouts.profile,
-        method: 'PUT',
-        body: userData,
-        headers: {
-          Authorization: `Bearer ${localStorage.getItem('accessToken')}`,
-        },
-      }),
-    }),
-  }),
-})
-
-export const { useGetProfileQuery, useUpdateProfileMutation } = profileApi
-
-export default profileApi
-
-// // Implement the server-side rendering for profile data
-// export async function getServerSideProps() {
-//   const { data: profileData } = await useGetProfileQuery()
-//   return {
-//     props: {
-//       profileData,
-//     },
-//   }
-// }
+export const profileApi_ = {
+  getProfile() {
+    return instance.get<AxiosResponse<ProfileData>>(userRouts.profile).then(({ data }) => data)
+  },
+}
